--- account/urls.py
@@ -1,8 +1,0 @@
-from django.urls import path
-from .views import UserDetailAPIView,RegisterUserAPIView
-
-
-urlpatterns = [
-  path("get-details",UserDetailAPIView.as_view()),
-  path('register',RegisterUserAPIView.as_view()),
-]
