+++ team66/urls.py
@@ -19,6 +19,6 @@ from rest_framework.authtoken import views
 
 urlpatterns = [
   path('admin/', admin.site.urls),
-  path('',include('account.urls')),
+  path('',include('apps.account.urls')),
   path('api-token-auth', views.obtain_auth_token)
 ]
